fix: remove metrics object no longer needed & cleanup
